refactor ray worker pipeline
